Moved cli button to title
