+++ dashboard/src/main/home/cluster-dashboard/expanded-chart/jobs/JobResource.tsx
@@ -290,12 +290,6 @@ export default class JobResource extends Component<PropsType, StateType> {
               <CommandString>{commandString}</CommandString>
               {this.renderStatus()}
               <MaterialIconTray disabled={false}>
-                <CLIModalIcon
-                  onClick={(e) => {
-                    e.preventDefault();
-                    this.setState({ showConnectionModal: true });
-                  }}
-                />
                 {this.renderStopButton()}
                 {!this.props.readOnly && (
                   <i
@@ -319,11 +313,6 @@ export default class JobResource extends Component<PropsType, StateType> {
           </MainRow>
           {this.renderLogsSection()}
         </StyledJob>
-        <ConnectToJobInstructionsModal
-          show={this.state.showConnectionModal}
-          onClose={() => this.setState({ showConnectionModal: false })}
-          job={this.props.job}
-        />
       </>
     );
   }
@@ -403,19 +392,6 @@ const Status = styled.div<{ color: string }>`
 const Icon = styled.img`
   width: 30px;
   margin-right: 18px;
-`;
-
-const CLIModalIcon = styled(CommandLineIcon)`
-  border-radius: 20px;
-  width: 26px;
-  padding: 5px;
-  margin: 0 5px;
-  :hover {
-    background: #ffffff11;
-  }
-  > path {
-    fill: #ffffff44;
-  }
 `;
 
 const Flex = styled.div`
